==> task_2/1.kyc_preprocessing.ipynb <==
from pathlib import Path
import pandas as pd
import numpy as np
import pyarrow
import json

# KYC Preprocessing

The purpose of this notebook is to add occupation-specific indicators to the KYC data. Indicators are binary flags representing whether or not a client's occupation is part of a specific subset of occupations. These subsets were determined using Fintrac and financial crime recommendations. Mainly:

Per the [Fintrac Operational Alert](https://fintrac-canafe.canada.ca/intel/operation/oai-wildlife-eng), one indicator is "An individual is the owner, operator, employee or associated with an industry that could be used to facilitate illegal wildlife trade (e.g., import/export of goods, fisheries wholesaler, pet store, freight company, animal control)." (**G**)

The [Financial Crime Academy](https://financialcrimeacademy.org/wildlife-trade-risk-indicators-financial/?fbclid=IwAR1XSw09Vtl4mjOOQj_eTFuqZ_GKqM-SPsCJwQKcyFb-XWU4O6nO8zBo3JU) adds that another "indicator relates to activity involving politically exposed persons and wealthy businessmen/women, particularly those with environmental, game, or forestry oversight or environmental or wildlife-related businesses." (**G**)

We have therefore come up with the following indicators based on KYC occupation data:
- `occ_wealth` 
    - binary 
    - 1 if the client's occupation involves frequent exposure to wealthy people, 0 otherwise.
- `occ_animal` 
    - binary 
    - 1 if the client's occupation involves working with animals, 0 otherwise.
- `occ_int` 
    - binary 
    - 1 if the client works international trade, 0 otherwise.
- `occ_shipping`
    - binary
    - 1 if the client works in shipping/postal/cargo services
    - *there are no examples of this in our data*

## Merging Occupation and Customer Data

# Adding the occupation-based indicators to raw data
DATAPATH = Path('../data/')
raw = pd.read_csv(DATAPATH / 'raw' / 'kyc.csv')
occ = pd.read_csv(DATAPATH / 'processed' / 'occupation_list.csv')

merged = raw.merge(occ, 'left', on='occupation')

# Re-order columns
cols = list(merged.columns)
cols.append(cols.pop(cols.index('label')))
merged = merged[cols]

## Getting All Customers

## Merging Task 3 Data

kyc_df = merged.copy()

# with open('../data/processed/names_metadata.json', 'r') as f:
#     names = json.load(f)

# kyc_df.loc[kyc_df['cust_id'].str.contains('CUST82758793')]

# kyc_df['named_trafficker']=0.0
# kyc_df['parsed_name']=kyc_df['name'].str.lower()

# for name, data in names.items():

#     score = float(data['case_name_score'])
    
#     #get number of matches
#     l = len(kyc_df.loc[kyc_df['parsed_name'].str.contains(name)])
    
#     #distribute score
#     kyc_df.loc[kyc_df['parsed_name'].str.contains(name), 'named_trafficker'] = score/(100*l)    
    
# kyc_df.drop(columns=['parsed_name'], inplace=True)

# Export
display(kyc_df.sample(3))
kyc_df.to_parquet(DATAPATH / 'processed' / 'kyc.parquet', index=False)

## Merging KYC and Transactional Data

kyc_df = merged.copy()
kyc_df = kyc_df.drop(['occupation', 'gender', 'name', 'age', 'tenure'], axis=1)

# Wire Transfer
wire_df = pd.read_csv(DATAPATH / 'raw' / 'wire.csv')
wire_df = wire_df.merge(kyc_df.add_suffix('_receiver', axis=1), on='cust_id_receiver', how='left')
wire_df = wire_df.merge(kyc_df.add_suffix('_sender', axis=1), on='cust_id_sender', how='left')
wire_df.to_parquet(DATAPATH / 'processed' / 'wire.parquet', index=False)
display(wire_df.sample(3))

# Emt
emt_df = pd.read_csv(DATAPATH / 'raw' / 'emt.csv')
emt_df = emt_df.merge(kyc_df.add_suffix('_receiver', axis=1), on='cust_id_receiver', how='left')
emt_df = emt_df.merge(kyc_df.add_suffix('_sender', axis=1), on='cust_id_sender', how='left')
emt_df.to_parquet(DATAPATH / 'processed' / 'emt.parquet', index=False)
display(emt_df.sample(3))

# Cash
cash_df = pd.read_csv(DATAPATH / 'raw' / 'cash.csv')
cash_df = cash_df.merge(kyc_df, on='cust_id', how='left')
cash_df.to_parquet(DATAPATH / 'processed' / 'cash.parquet', index=False)
display(cash_df.sample(3))
